+++ src/test/java/in/sakthi/TestDiscount.java
@@ -1,14 +1,15 @@
 package in.sakthi;
 
 import in.sakthi.model.Discount;
+import in.sakthi.service.DiscountDate;
 import in.sakthi.service.DiscountService;
 
 public class TestDiscount {
 	public static void main(String[] args) {
 		String channelName = "Hotstar";
-		String validity = "Yearly";
+		String validity = "Monthly";
 		Discount list = new Discount(channelName, validity);
 		DiscountService.discount(list);
-		DiscountService.date(validity);
+		DiscountDate.date(validity);
 	}
 }
